Ensure offset calculation is 64-bit safe

+++ code/botlib/be_ai_weap.c
@@ -39,8 +39,8 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 //#define DEBUG_AI_WEAP
 
 // structure field offsets
-#define WEAPON_OFS(x) (int)&(((weaponinfo_t*)0)->x)
-#define PROJECTILE_OFS(x) (int)&(((projectileinfo_t*)0)->x)
+#define WEAPON_OFS(x) ((size_t) & (((weaponinfo_t*)0)->x))
+#define PROJECTILE_OFS(x) ((size_t) & (((projectileinfo_t*)0)->x))
 
 // weapon definition
 static fielddef_t weaponinfo_fields[] = {
